# specs/001-tandem-slack-bot/contracts/api.yaml
openapi: 3.0.3
info:
  title: Tandem Slack Bot API
  description: API for managing tasks detected from Slack and scheduling them in Google Calendar
  version: 1.0.0
  contact:
    name: Tandem Team
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://api-dev.tandem.app/api
    description: Development environment
  - url: https://api.tandem.app/api
    description: Production environment
security:
  - BearerAuth: []
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      summary: Get user's tasks
      description: Retrieve tasks with optional filtering
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, scheduled, completed, dismissed]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: Get specific task
      description: Retrieve detailed task information
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags: [Tasks]
      summary: Update task
      description: Update task details before or after scheduling
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
        estimatedDuration:
          type: integer
          description: Duration in minutes
        importance:
          type: string
          enum: [Low, Medium, High]
        derivedUrgency:
          type: string
          enum: [Low, Medium, High]
        status:
          type: string
          enum: [pending, confirmed, scheduled, completed, dismissed]
        calendarEvent:
          $ref: '#/components/schemas/CalendarEvent'
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, dueDate, estimatedDuration, importance, derivedUrgency, status, createdAt, updatedAt]
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
        estimatedDuration:
          type: integer
          minimum: 5
          maximum: 480
        importance:
          type: string
          enum: [Low, Medium, High]
    TaskList:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [tasks, total, limit, offset]

    # Calendar event schemas
    CalendarEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        googleEventId:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        isActive:
          type: boolean
      required: [id, googleEventId, startTime, endTime, isActive]

    # Preferences schemas
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              nullable: true
          required: [code, message]
      required: [error]
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...